refactor!: Move `StringSortedSetConverter` to `ort-utils`
This makes `common-utils` Jackson-free, which is a step towards being
able to make it a multi-platform project.
